imageupload/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.0
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
wcwidth==0.1.7
zipp==0.6.0
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.2.0
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
wcwidth==0.1.7
zipp==0.6.0
